+++ BadBits.Engine/Model/Attribs.cs
@@ -4,9 +4,9 @@ namespace BadBits.Engine.Model
 {
     public class vertexPosition
     {
-        public float x { get; set; }
-        public float y { get; set; }
-        public float z { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
+        public double z { get; set; }
     }
 
     public class TextureAttribs
